dit637/07/modelserver-fastapi/requirements.txt
fastapi
pydantic
transformers
uvicorn
scikit-learn
python-dotenv
mlflow
pandas
